+++ packages/components/src/index.ts
@@ -1,3 +1,4 @@
 export * from './components/button/index.js';
 export * from './components/checkbox/index.js';
 export * from './components/radio/index.js';
+export * from './components/toggle/index.js';